Implement DeleteCashDividend
When the dividend is deleted, the associated Events are removed from the
CashDividend as well as the Accounts.

The menu command "DCD" uses a new style where the ID of the dividend to be
deleted is given as a parameter; it is not entered separately. I want to move
to this style because it feels more natural. I will do all new commands this
way, and redo existing commands as time allows.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>