Update StudentModelContext to my new way of doing things
For the DbContext I am using two constructors: one that takes the path and one
that uses DbContextOptions, which allows me to use a design-time factory to
specify a different database file.

Also fixed a bug in the student sample where John was using the first Grade in
the list returned by the database. This fails when the database is new, since
there are no Grades. The fix is to use the new Grade that was just created.